+++ app/Http/Requests/Product/UpdateProductRequest.php
@@ -22,6 +22,7 @@ class UpdateProductRequest extends FormRequest
         return [
             'name' => ['required', 'string', 'max:255'],
             'description' => ['nullable', 'string'],
+            'brand_id' => ['nullable', 'string', 'exists:tenant.brands,uuid'],
             'product_type_id' => ['required', 'string', 'exists:tenant.product_types,uuid'],
             'currency_id' => ['nullable', 'string', 'exists:tenant.currencies,uuid'],
             'is_active' => ['boolean'],