+++ app/Models/Product.php
@@ -25,6 +25,11 @@ class Product extends Model
         ];
     }
 
+    public function brand(): BelongsTo
+    {
+        return $this->belongsTo(Brand::class);
+    }
+
     public function productType(): BelongsTo
     {
         return $this->belongsTo(ProductType::class);